+++ routes/category.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const { getCategory } = require('../controller/categoryController');
-const ensureAuthorization = require('../middleware/decodedJWT');
+const ensureAuthorization = require('../middleware/ensureAuthorization');
 
 router.get('/', ensureAuthorization(false), getCategory);
