+++ routes/likes.js
@@ -3,7 +3,7 @@ const router = express.Router();
 
 const { postLike, deleteLike } = require('../controller/likesController');
 const { validateLikes } = require('../validators/likes');
-const ensureAuthorization = require('../middleware/decodedJWT');
+const ensureAuthorization = require('../middleware/ensureAuthorization');
 
 router
   .route('/:book_id')
